+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### v0.2.0
 - 🍏 add new `BaseElement`
 - 🍏 add new `loot` drag-and-drop system
+- 🍏 add new tool `dom.events` for attaching dom events
 - 🟥 rework view rendering syntax
     - 😡 old bad
         ```ts